+++ Deployment/tooling/includes/run.fns.sh
@@ -6,11 +6,24 @@
 function runInDocker() {
   h2 "Run incremental build in a docker environment"
   RUNTIME_IMAGE="${IMAGENAME}-runtime:latest" docker compose \
-    -f "${PROJECT_DIR}/Deployment/docker-compose.yml" \
+    -f "${PROJECT_DIR}/Deployment/build-compose.yml" \
     --env-file "${SCRIPT_DIR}/.env" \
     up \
     --exit-code-from run \
     run
+}
+
+# Run official Uitsmijter in docker compose 
+function runInDockerProduction() {
+  h2 "Run Uitsmijter in production environment"
+  local IMAGENAME="${1:-${IMAGENAME}}"
+  local TAG="${2:-${TAG}}"
+  buildReleaseIfNotPresent "${IMAGENAME}:${TAG}"
+  IMAGENAME=${IMAGENAME} TAG=${TAG} \
+    docker compose \
+    -f "${PROJECT_DIR}/Deployment/Docker/docker-compose.yml" \
+    --env-file "${PROJECT_DIR}/Deployment/Docker/.env" \
+    up
 }
 
 # Run a release or prerelease in kind
@@ -22,7 +35,7 @@ function runInKubernetesInDocker() {
   kindStartCluster "${TAG}"
   echo ''
   echo 'Cluster is running.'
-  # Keep in sync with docker-compose.yml
+  # Keep in sync with build-compose.yml
   echo 'Add the following to your local /etc/hosts file.'
   echo ''
   for host in ${TEST_HOSTS}; do