+++ CHANGELOG.md
@@ -16,8 +16,10 @@
 
 - Update: to Swift 5.9.2
 - Update buildbox to 2.3.0
+- Update Vapor to 4.106.1
 
 - Improvement: Tooling command code use its own build folder
+- Improvement: Tooling command test can have an optional filter
 - Improvement: Rename "Littleletetr" to "Ham" to align the documentation
 
 # 0.9.5